Maintain release notes with GitHub Actions
This commit sets up a GitHub Actions workflow that uses Release Drafter
to automatically maintain release notes on every push.

In practice this means that contributors no longer have to manually
update `Changelog.md`, which is something that is easy to forget and
often introduces tedious merge conflicts.

Instead, Release Drafter automatically adds the title of the PR to the
GitHub release notes to a draft release, crediting the author of the PR,
and linking to the PR number. Release Drafter furthermore organizes the
release notes into sections according to the labels assigned to the PRs:

- ⚠️ Breaking
- 🐛 Bug Fix
- 📚 Docs
- ✨ Feature
- 🏠 Housekeeping