Add Gemfile.lock to gitignore (#18)
We use a matrix in CI to test the project against various versions of
Ruby. It is possible that different versions will have different
dependency resolutions, hence different lockfiles.

Instead of pinning the project to a specific version of Ruby by
committing that version's lockfile, I prefer to keep the lockfile out of
git.